# Memory and performance optimization [API-422]