--- a/events.go
+++ b/events.go
@@ -60,14 +60,14 @@ const (
 )
 
 type EntryNotified struct {
-	EventType               int32
-	MapName                 string
+	MergingValue            interface{}
 	Member                  cluster.Member
 	Key                     interface{}
 	Value                   interface{}
 	OldValue                interface{}
-	MergingValue            interface{}
+	MapName                 string
 	NumberOfAffectedEntries int
+	EventType               int32
 }
 
 func (e *EntryNotified) EventName() string {
@@ -126,10 +126,10 @@ func newLifecycleStateChanged(state LifecycleState) *LifecycleStateChanged {
 }
 
 type MessagePublished struct {
-	TopicName   string
-	Value       interface{}
 	PublishTime time.Time
+	Value       interface{}
 	Member      cluster.Member
+	TopicName   string
 }
 
 func (m *MessagePublished) EventName() string {
@@ -148,9 +148,9 @@ func newMessagePublished(name string, value interface{}, publishTime time.Time, 
 type QueueItemNotifiedHandler func(event *QueueItemNotified)
 
 type QueueItemNotified struct {
-	QueueName string
 	Value     interface{}
 	Member    cluster.Member
+	QueueName string
 	EventType int32
 }