--- a/go.mod
+++ b/go.mod
@@ -6,6 +6,5 @@ require (
 	github.com/apache/thrift v0.14.1
 	github.com/stretchr/testify v1.6.1
 	go.uber.org/goleak v1.1.10
-	golang.org/x/lint v0.0.0-20201208152925-83fdc39ff7b5 // indirect
 	golang.org/x/tools v0.1.0 // indirect
 )