--- a/internal/cluster/client_message.go
+++ b/internal/cluster/client_message.go
@@ -23,7 +23,7 @@ import (
 	"github.com/hazelcast/hazelcast-go-client/internal/proto"
 )
 
-const messageBufferSize = 1024
+const messageBufferSize = 64
 
 type clientMessageReader struct {
 	src                *bytes.Buffer
@@ -31,7 +31,6 @@ type clientMessageReader struct {
 	readHeader         bool
 	currentFrameLength uint32
 	currentFlags       uint16
-	remainingCap       int
 }
 
 func newClientMessageReader() *clientMessageReader {
@@ -47,7 +46,7 @@ func (c *clientMessageReader) Append(buf []byte) {
 func (c *clientMessageReader) Read() *proto.ClientMessage {
 	for {
 		if c.readFrame() {
-			if c.clientMessage.EndFrame.IsFinalFrame() {
+			if c.clientMessage.HasFinalFrame() {
 				return c.clientMessage
 			}
 		} else {