--- a/internal/hzerror/server_error.go
+++ b/internal/hzerror/server_error.go
@@ -17,12 +17,12 @@
 package hzerror
 
 type ServerError struct {
-	errorCode      int32
 	className      string
 	message        string
-	stackTrace     []StackTraceElement
-	causeErrorCode int32
 	causeClassName string
+	stackTrace     []StackTraceElement
+	errorCode      int32
+	causeErrorCode int32
 }
 
 // NewServerErrorImpl