This PR:
* Fixes writing payloads larger than 8 KBs.
* Refactors serialization, especially serializing byte arrays and strings is much more performant.
* Internal serialization functions panic instead of returning errors
* Updated portable and identified serialization interfaces.
* Makes some of the public serialization types internal
* Refactors frames in messages, uses a frame slice instead of `frame.next` 
* Adds some benchmarks
* Optimizes structs so their memory layout is optimal
* Fixes non-retryable errors
* Enables socket deadlines
* Increases connection write buffer size, decreases initial read buffer size
* Added missing connection manager event handler for closed connections
* Removes unnecessary defers.