@puzpuzpuz 2021-05-14
The PR description seems to lack details on the optimizations, so I tried running some benchmarks on my machine.

Environment: single member run with `docker run --net=host hazelcast/hazelcast:4.2`, go1.16.2 linux/amd64, `uname -a`: `Linux apechkurov-laptop 5.4.0-73-generic #82-Ubuntu SMP Wed Apr 14 17:39:42 UTC 2021 x86_64 x86_64 x86_64 GNU/Linux`, i5-8400h

1. `go test -bench BenchmarkMap_SetSameKeyValue -benchmem ./benchmarks`

`master`:
```
BenchmarkMap_SetSameKeyValue/Smart_Client-8         	   27590	     43026 ns/op	    2336 B/op	      44 allocs/op
BenchmarkMap_SetSameKeyValue/Non-Smart_Client-8     	   29433	     40394 ns/op	    2184 B/op	      38 allocs/op
```

This PR:
```
BenchmarkMap_SetSameKeyValue/Smart_Client-8         	   26536	     48199 ns/op	    1385 B/op	      41 allocs/op
BenchmarkMap_SetSameKeyValue/Non-Smart_Client-8     	   25939	     44491 ns/op	    1232 B/op	      35 allocs/op
```

Allocation rate decreased significantly, but the throughput didn't improve.

2. `go test -bench BenchmarkMap_SetParallel -benchmem ./benchmarks` (didn't include `BenchmarkMap_SetParallelLargePayload_128KB` benchmark since it crashes on `master`)

`master`:
```
BenchmarkMap_SetParallel/Smart_Client-8        	   93914	     12201 ns/op	    2337 B/op	      44 allocs/op
BenchmarkMap_SetParallel/Non-Smart_Client-8    	   95451	     14168 ns/op	    2184 B/op	      38 allocs/op
```

This PR:
```
BenchmarkMap_SetParallel/Smart_Client-8        	   95246	     12233 ns/op	    1388 B/op	      41 allocs/op
BenchmarkMap_SetParallel/Non-Smart_Client-8    	   88525	     12190 ns/op	    1235 B/op	      35 allocs/op
```

That's basically the same scenario as the previous one, but run on all cores, so no surprise that the outcome is that same.

3. `go test -bench BenchmarkMap_SetGetDifferentKeyValue -benchmem ./benchmarks`

`master`:
```
BenchmarkMap_SetGetDifferentKeyValue/Smart_Client-8         	   13537	     87804 ns/op	    4742 B/op	      90 allocs/op
BenchmarkMap_SetGetDifferentKeyValue/Non-Smart_Client-8     	   14084	     85108 ns/op	    4438 B/op	      78 allocs/op
```

This PR:
```
BenchmarkMap_SetGetDifferentKeyValue/Smart_Client-8         	   13538	     88868 ns/op	    2811 B/op	      89 allocs/op
BenchmarkMap_SetGetDifferentKeyValue/Non-Smart_Client-8     	   14180	     84377 ns/op	    2507 B/op	      77 allocs/op
```

Once again, allocation rate is now lower, the number of allocation is slightly lower, but the throughput is on the same level.

Looks like this PR aims to optimize memory consumption and, probably, improves things for large payloads (didn't check that yet), rather than general performance. Going to continue with the review now.