@yuce 2021-05-14
@puzpuzpuz I've updated the description with the changes. This PR fixes  writing payloads larger than 8KBs, so 128KB benchmark doesn't run against `master`. I'll create a temporary PR which fixes that on master, so you can run the benchmark against `master`. This PR significantly improves `Put`, `Get` like calls for byte arrays and strings. In my tests it was more than 100x faster. It improves `Set` against more than about 1000x against v3 and `Get` less-dramatic, like 10x (if I remember correctly, will check again). The improvement comes from pre-allocating the byte array in the output, instead of making the output 1 byte larger for every byte in the byte array! I'll post some gists shortly.