@yuce 2021-05-14
I've ran the code in the following gists with a 512 KB payload (single member), so the results are more dramatic.
* v3: https://gist.github.com/yuce/81b6e5a5ccac3875478e3fcc9229257e : Set: 31288 ms, Get: 56 ms
* this pr: https://gist.github.com/yuce/edec1b18040f68e506ed4268d9b7d693 Set: 1 ms, Get: 58 ms

I've ran the code above a few times to warm the Hz instance. Although definitely not-scientific, it shows the 4 orders of magnitude between v3 and v1 `Set`.