@yuce 2021-05-14
> 
> 
> > I've ran the code in the following gists with a 512 KB payload (single member), so the results are more dramatic.
> 
> That's basically the same story as with Node.js client pre-v3 -> v3 -> v4. String and byte array serialization was a bottleneck there and the throughput dropped exponentially for larger payloads. That's certainly great to see an improvement for large payloads. On the other hand, the "recommended" (and, certainly, typical) payload size is in the 1-10KB range.

10KB payload Set with v3 is in 14-20 ms range and it's 0 ms with this PR, so I guess that makes this one infinitely faster :D