@puzpuzpuz 2021-05-14
> 10KB payload Set with v3 is in 14-20 ms range and it's 0 ms with this PR, so I guess that makes this one infinitely faster :D

Yeah, old serialization of strings (and byte arrays in some clients) was far from being optimal. Standard libraries are much better than the custom implementations.

See:
https://hazelcast.com/blog/our-journey-to-a-high-performance-node-js-library/
https://github.com/hazelcast/hazelcast-nodejs-client/pull/473
https://github.com/hazelcast/hazelcast/pull/15238